Bumps deployment target to 9.0, replaces deprecated constants & fixes warnings

+++ Project/UISS/UISSErrorsViewController.m
@@ -35,7 +35,7 @@
 
 - (void)update {
     if (self.errors.count) {
-        self.tabBarItem.badgeValue = [NSString stringWithFormat:@"%d", self.errors.count];
+        self.tabBarItem.badgeValue = [NSString stringWithFormat:@"%lu", (unsigned long) self.errors.count];
     } else {
         self.tabBarItem.badgeValue = nil;
     }
